Tests/LibWeb/Text/expected/WebVTT/VTTCue-settings.txt

line: "auto"
line: 42
line: "auto"
position: "auto"
position: 42
position: "auto"
lineAlign: "start"
lineAlign: "center"
lineAlign: "end"
positionAlign: "auto"
positionAlign: "line-left"
positionAlign: "center"
positionAlign: "line-right"
align: "start"
align: "center"
align: "end"
align: "left"
align: "right"
vertical: ""
vertical: "rl"
vertical: "lr"
